Move goals frontend into feature module
