+++ apps/frontend/src/features/goals/hooks/use-goal-detail.ts
